factor out  unit tests
