+++ mysql/tests/common.py
@@ -11,7 +11,7 @@ HERE = os.path.dirname(os.path.abspath(__file__))
 ROOT = os.path.dirname(os.path.dirname(HERE))
 TESTS_HELPER_DIR = os.path.join(ROOT, 'datadog_checks_tests_helper')
 
-MYSQL_VERSION_PARSED = parse_version(os.getenv('MYSQL_VERSION'))
+MYSQL_VERSION_PARSED = parse_version(os.getenv('MYSQL_VERSION'), '')
 
 CHECK_NAME = 'mysql'
